# src/main/resources/openapi-checkout-v1.yaml
openapi: 3.0.0
info:
  title: Shopping Cart Checkout API
  version: 1.0.0
  description: API for handling the checkout process.
servers:
  - url: /api/v1
    description: API base path
tags:
  - name: Checkout Process
    description: Operations to handle the checkout process.
paths:
  /checkout:
    post:
      summary: Initiate checkout with a full shopping cart
      operationId: initiateCheckout
      tags:
        - Checkout
      description: Receives a complete shopping cart entity to initiate the checkout process.
      requestBody:
        description: The complete shopping cart entity to be processed for checkout.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShoppingCart'
            example:
              cartId: "abcd1234-efgh-5678-ijkl-901234567890"
              userId: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
              items:
                - productId: "E12345678"
                  productName: "Basic License"
                  quantity: 1
                  price: 1200.00
                - productId: "E87654321"
                  productName: "Premium License"
                  quantity: 2
                  price: 25.00
              totalItems: 3
              totalAmount: 1250.00
      responses:
        '200':
          description: Shopping Cart published successfully
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Failed to publish Shopping Cart
          content:
            text/plain:
              schema:
                type: string
  /cart/{cartId}:
    get:
      summary: Retrieve a shopping cart by its ID
      operationId: getCartById
      tags:
        - Checkout
      description: Fetches the details of a specific shopping cart using its unique identifier.
      parameters:
        - name: cartId
          in: path
          description: Unique identifier of the shopping cart to retrieve.
          required: true
          schema:
            type: string
            format: uuid
          example: "abcd1234-efgh-5678-ijkl-901234567890"
      responses:
        '200':
          description: Shopping cart found and returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShoppingCart'
        '404':
          description: Shopping cart with the specified ID not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "CART_NOT_FOUND"
                message: "Shopping cart with ID 'abcd1234-efgh-5678-ijkl-901234567890' not found."
        '500':
          description: Internal server error (e.g., failed to retrieve cart data).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ShoppingCart:
      type: object
      description: Represents a shopping cart.
      required:
        - cartId
        - items
        - totalItems
        - totalAmount
      properties:
        cartId:
          type: string
          format: uuid
          description: Unique identifier for this shopping cart. This will often serve as the sessionId for checkout correlation.
          example: "abcd1234-efgh-5678-ijkl-901234567890"
        userId:
          type: string
          description: (Optional) ID of the user associated with the cart.
          example: "V12345678"
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
          description: List of items currently in the cart.
          minItems: 1
        totalItems:
          type: integer
          description: Total number of distinct items in the cart.
          minimum: 0
        totalAmount:
          type: number
          format: float
          description: Total monetary value of all items in the cart.
          minimum: 0
    CartItem:
      type: object
      description: Represents a single item in the shopping cart.
      required:
        - productId
        - productName
        - quantity
        - price
      properties:
        productId:
          type: string
          description: Unique identifier of the product.
          example: "E12345678"
        productName:
          type: string
          description: Name of the product.
          example: "Basic License"
        quantity:
          type: integer
          minimum: 1
          description: Quantity of the product in the cart.
          example: 1
        price:
          type: number
          format: float
          description: Unit price of the product at the time it was added to the cart.
          minimum: 0
          example: 1200.00
    ErrorResponse:
      type: object
      description: Standard error response format.
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: A unique error code.
          example: "INVALID_CART_DATA"
        message:
          type: string
          description: A human-readable error message.
          example: "The provided cart data is invalid or empty."